spliting hpp/cpp for bindings
left multipatch untouched to avoid merge conflict with #194
minor class layout change spline reader
