+++ CMakeLists.txt
@@ -91,12 +91,10 @@ if(SPLINEPY_COMPILE_PYTHON)
   if(splinepy_FOUND)
     message("Found installed splinepy, building only python module")
     add_subdirectory(third_party/pybind11)
-    #add_subdirectory(src/py)
-    add_subdirectory(include/splinepy/py)
+    add_subdirectory(src/py)
     return()
   endif()
 endif()
 
 add_subdirectory(third_party)
 add_subdirectory(src)
-add_subdirectory(include/splinepy/py)
