--- include/splinepy/py/init/coordinate_pointers.cpp
@@ -1,7 +1,0 @@
-#include "splinepy/py/py_coordinate_pointers.hpp"
-
-namespace splinepy::py::init {
-
-void init_coordinate_pointers(py::module_& m) { add_coordinate_pointers(m); }
-
-} // namespace splinepy::py::init
