--- include/splinepy/py/init/fitting.cpp
@@ -1,7 +1,0 @@
-#include <splinepy/py/py_fitting.hpp>
-
-namespace splinepy::py::init {
-
-void init_fitting(py::module_& m) { add_fitting(m); }
-
-} // namespace splinepy::py::init
