--- include/splinepy/py/init/multi_patch.cpp
@@ -1,7 +1,0 @@
-#include <splinepy/py/py_multi_patch.hpp>
-
-namespace splinepy::py::init {
-
-void init_multi_patch(py::module_& m) { add_multi_patch(m); }
-
-} // namespace splinepy::py::init
