--- include/splinepy/py/init/reader.cpp
@@ -1,7 +1,0 @@
-#include <splinepy/py/py_spline_reader.hpp>
-
-namespace splinepy::py::init {
-
-void init_reader(py::module_& m) { add_spline_reader(m); }
-
-} // namespace splinepy::py::init
